QOL Updates
Removed double spaces in text descriptions so they're shorter.
Added sustain stat to all artifacts that boost that stat by 3+
Modified drop rates for bows to offset how few there are compared to other weapon types
Removed breakage on arrows as a bandaid - want to ultimately only remove breakage on mithril arrows.
Changed stealth so for all classes so that it improves slowly with CLVL.
Changed inventory size for player and home - greatly increasing home size so you don't have to discard so many artifacts and can stockpile good ammo for a different missile launcher
changed drop rate on some books so that you can gain access to spells a bit closer to the actual level when you would be able to learn them.
Changed probability for some monsters and some capabilities to remove the more silly ones that can't be resisted like TIME